+++ src/router/index.js
@@ -1,6 +1,7 @@
 import { createRouter, createWebHashHistory } from 'vue-router'
 import Home from '../views/Home.vue'
 import GiphyView from '@/views/GiphyView.vue'
+import MovieView from '@/views/MovieView.vue'
 import PageNotFoundView from '@/views/PageNotFoundView.vue'
 
 const routes = [
@@ -15,16 +16,13 @@ const routes = [
     component: GiphyView
   },
   {
+    path: '/movie',
+    name: 'Movie',
+    component: MovieView
+  },
+  {
     path: '/:catchAll(.*)*',
     component: PageNotFoundView
-  }, 
-  {
-    path: '/about',
-    name: 'About',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
   }
 ]