+++ src/views/MovieView.vue
@@ -1,0 +1,11 @@
+<template>
+    <div>
+        Movie View
+    </div>
+</template>
+
+<<script>
+export default {
+    
+}
+</script>